+++ .pre-commit-config.yaml
@@ -1,3 +1,5 @@
+# exclude: ^site/|^test_html
+
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.3.0
