+++ README.md
@@ -6,28 +6,31 @@ This repo contains a set of economics reports stored as jupyter notebooks. The n
 
 - Inflation - reports the change in prices overtime
 
-## Setup
+## Setup for development / contributing
 
 1. Install a version of python with [pyenv](https://github.com/pyenv/pyenv)
 2. Install [poetry](https://python-poetry.org/)
-3. Clone the repo
-3. Run `poetry install`
-4. Run `poetry shell`
+3. Fork the repo
+4. Clone the forked repo
+5. Run `poetry install --with dev`
+6. Run `poetry shell`
+7. Setup pre-commit hooks by running `pre-commit install`
 
-## How-to
+## How to generate reports
 
 The reports can be generated with a `make report` command. The make command will execute the notebook with [papermill](https://papermill.readthedocs.io/en/latest/) and then use [nbconvert](https://nbconvert.readthedocs.io/en/latest/) to convert the notebook into an html file.
 
-Use the example command below to generate a report. You can view the html report by opening the html file at `/output/html/{report}_{run_date}.html`, or the jupyter notebook at `/output/notebooks/{report}_{run_date}.ipynb`
+For example, you can generate the cpi report with the make command below.
 
 ```bash
-make report report={name_of_report}
+make report report=cpi
 ```
 
+This command run two steps
+    1. Papermill is used to run `cpi.ipynb` file in the `notebooks` directory and stores the output notebook at `site/source/notebooks`
+    2. nbconvert converts the output notebook into an html file at `site/source/html`
 
-## Setup for development
+## Guide for building notebooks
 
-1. Make a fork of the repo
-2. Clone your repo to your machine
-3. Install the development dependencies with `poetry install --with dev`
-4. Setup pre-commit hooks by running `pre-commit install`
+- If you want the cell excluded from the html output, add the tag `remove_cell`
+- If you want to remove the cell code, but not the output, add the tag `remove_input`
