--- econ_reports_site/source/_templates/layout.html
@@ -1,8 +1,0 @@
-{# Import the theme's layout. #}
-{% extends '!layout.html' %}
-
-{%- block extrahead %}
-{# Add custom things to the head HTML tag #}
-{# Call the parent block #}
-{{ super() }}
-{%- endblock %}
