--- econ_reports_site/source/index.rst
@@ -1,17 +1,0 @@
-.. econ_reports documentation master file, created by
-   sphinx-quickstart on Thu Oct 27 07:46:39 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to Econ Reports!
-========================================
-
-.. toctree::
-   :caption: Reports
-   :maxdepth: 1
-
-   CPI <notebooks/reports/cpi_report.ipynb>
-
-.. toctree::
-   :caption: Analysis:
-   :maxdepth: 2
