--- econ_reports_utils/pyproject.toml
@@ -1,8 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "econ_reports_utils"
-version = "0.1.0"
-requires-python = ">=3.7"
