+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Economic reports with jupyter notebooks"
 authors = ["gw-moore <gwmoore92@gmail.com>"]
 license = "MIT"
 readme = "README.md"
-packages = [{include = "econ_reports"}]
+packages = [{include = "econ_reports_utils"}]
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.11"
@@ -20,10 +20,6 @@ rich = "^12.6.0"
 altair = "^4.2.0"
 seaborn = "^0.12.0"
 statsmodels = "^0.13.2"
-Sphinx = "^5.3.0"
-sphinx-material = "^0.0.35"
-myst-parser = "^0.18.1"
-nbsphinx = "^0.8.9"
 
 [tool.poetry.group.dev.dependencies]
 black = {extras = ["jupyter"], version = "^22.8.0"}
